[Refactor]: 챌린지 리뷰 서비스 분리
