+++ zzansuni-api-server/app/src/main/java/org/haedal/zzansuni/infrastructure/userchallenge/review/ChallengeReviewReaderImpl.java
@@ -1,11 +1,6 @@
 package org.haedal.zzansuni.infrastructure.userchallenge.review;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.NoSuchElementException;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.haedal.zzansuni.domain.userchallenge.ChallengeReview;
 import org.haedal.zzansuni.domain.userchallenge.port.ChallengeReviewReader;
@@ -14,7 +9,9 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
 
-import static org.haedal.zzansuni.domain.userchallenge.review.QChallengeReview.challengeReview;
+import java.util.*;
+
+import static org.haedal.zzansuni.domain.userchallenge.QChallengeReview.challengeReview;
 
 
 @Component
